backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.2
pillow==10.1.0
firebase-admin==6.3.0
functions-framework==3.5.0
